app/src/main/java/com/dedesaepulloh/submissionbajp/ui/tvshow/TvShowViewModel.kt
package com.dedesaepulloh.submissionbajp.ui.tvshow

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.dedesaepulloh.submissionbajp.data.source.CatalogRepository
import com.dedesaepulloh.submissionbajp.data.source.local.entity.TvShowEntity
import com.dedesaepulloh.submissionbajp.vo.Resource

class TvShowViewModel(private val catalogRepository: CatalogRepository) : ViewModel() {
    fun getTvShow(): LiveData<Resource<PagedList<TvShowEntity>>> =
        catalogRepository.getTvShowPopular()
}